+++ changelog.md
@@ -4,6 +4,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v0.0.1 (unreleased)
 
+-   [build][fix] Export TypeScript definitions
 -   [update] Add initial TextInput styles and props
 -   [build][dev] Enable development releases
 -   [dev] Add wikimedia-ui theme